一般说只有你要Extend immutable type like str, int, unicode or tuple时才会写__new__ Mastering Object-oriented Python-1.pdf 第二章最后有例子 Python.3.Object.Oriented.Programming.Dusty.Phillips.2010.pdf的第6章也有个例子，是给一个dict增加feature:含有key的ordered list
就现实而言，你大概只有写元类才要__new__当那样做时记住第一能不用元类不用之第二用别人的惯用法去stackoverflow而不是要去自己理解太深

The __new__() method is auto-magically a static method. This is true without using the @staticmethod decorator. It doesn't use a self variable, as its job is to create the object that will eventually be assigned to the self variable.

Use __new__ when you need to control the creation of a new instance. Use __init__ when you need to control initialization of a new instance.
__new__ is the first step of instance creation. It's called first, and is responsible for returning a new instance of your class. In contrast, __init__ doesn't return anything; it's only responsible for initializing the instance after it's been created.
In general, you shouldn't need to override __new__ unless you're subclassing an immutable type like str, int, unicode or tuple.